Moved bilby_inference.py
