+++ bilby_inference/bilby_inference.py
@@ -1,0 +1,242 @@
+import os
+import subprocess as sp
+from collections import OrderedDict
+
+import corner
+import h5py
+import matplotlib.font_manager as font_manager
+import numpy as np
+from astropy.utils.data import download_file
+from bilby.core.prior import Uniform
+from comparitors import comparisons
+from cwinpy import HeterodynedData
+from cwinpy.pe import pe
+from lalinference import LALInferenceHDF5PosteriorSamplesDatasetName
+from lalinference.io import read_samples
+from matplotlib.lines import Line2D
+import pickle
+
+def pickler(path,obj):
+    """
+    Pickles an object and saves it to a chosen location
+
+    Parameters
+    ----------
+    path : str
+        Path of directory to save .pck file.
+    obj : Python object 
+        Object to pickle.
+
+    Returns
+    -------
+    None.
+
+    """
+    outfile = open(path,'wb')
+    pickle.dump(obj,outfile)
+    outfile.close()
+    print(path+" pickled")
+
+
+
+parcontent = """\
+PSRJ     J0123+3456
+RAJ      01:23:45.6789
+DECJ     34:56:54.321
+F0       567.89
+F1       -1.2e-12
+PEPOCH   56789
+H0       1.1e-25
+COSIOTA  0.01
+PSI      1.1
+PHI0     2.4
+"""
+
+injection_parameters = OrderedDict()
+injection_parameters["h0"] = 1.1e-25
+injection_parameters["phi0"] = 2.4
+injection_parameters["psi"] = 1.1
+injection_parameters["cosiota"] = 0.01
+
+detector = "H1"  # the detector to use
+asd = 1e-24  # noise amplitude spectral density
+
+
+label = "single_detector_software_injection_linear"
+outdir=""
+
+parfile = os.path.join(outdir, "{}.par".format(label))
+with open(parfile, "w") as fp:
+    fp.write(parcontent)
+
+times = np.linspace(1000000000.0, 1000086340.0, 1440)
+
+het = HeterodynedData(
+    times=times,
+    par=parfile,
+    injpar=parfile,
+    inject=True,
+    fakeasd=asd,
+    detector=detector,
+)
+
+# output the data
+hetfile = os.path.join(outdir, "{}_data.txt".format(label))
+het.write(hetfile)
+
+
+phi0range = [0.0, np.pi]
+psirange = [0.0, np.pi / 2.0]
+cosiotarange = [-1.0, 1.0]
+h0range = [0.0, 1e-23]
+
+# set prior for lalapps_pulsar_parameter_estimation_nested
+priorfile = os.path.join(outdir, "{}_prior.txt".format(label))
+priorcontent = """H0 uniform {} {}
+PHI0 uniform {} {}
+PSI uniform {} {}
+COSIOTA uniform {} {}
+"""
+with open(priorfile, "w") as fp:
+    fp.write(priorcontent.format(*(h0range + phi0range + psirange + cosiotarange)))
+
+# set prior for bilby
+priors = OrderedDict()
+priors["h0"] = Uniform(h0range[0], h0range[1], "h0", latex_label=r"$h_0$")
+priors["phi0"] = Uniform(
+    phi0range[0], phi0range[1], "phi0", latex_label=r"$\phi_0$", unit="rad"
+)
+priors["psi"] = Uniform(
+    psirange[0], psirange[1], "psi", latex_label=r"$\psi$", unit="rad"
+)
+priors["cosiota"] = Uniform(
+    cosiotarange[0], cosiotarange[1], "cosiota", latex_label=r"$\cos{\iota}$"
+)
+
+
+# run lalapps_pulsar_parameter_estimation_nested
+try:
+    execpath = os.environ["CONDA_PREFIX"]
+except KeyError:
+    raise KeyError(
+        "Please work in a conda environment with lalsuite and cwinpy installed"
+    )
+
+execpath = os.path.join(execpath, "bin")
+
+lppen = os.path.join(execpath, "lalapps_pulsar_parameter_estimation_nested")
+outfile = os.path.join(outdir, "{}_nest.hdf".format(label))
+
+n2p = os.path.join(execpath, "lalinference_nest2pos")
+# convert nested samples to posterior samples
+outpost = os.path.join(outdir, "{}_post.hdf".format(label))
+runcmd = " ".join([n2p, "-p", outpost, outfile])
+
+
+
+# get posterior samples
+post = read_samples(outpost, tablename=LALInferenceHDF5PosteriorSamplesDatasetName)
+lp = len(post["H0"])
+postsamples = np.zeros((lp, len(priors)))
+for i, p in enumerate(priors.keys()):
+    postsamples[:, i] = post[p.upper()]
+
+
+Nlive = 1024  # number of nested sampling live points
+
+
+try:    
+    infile = open("bilby_result.pkl",'rb')       #Try to load relevent posterior 
+    result = pickle.load(infile)
+    infile.close()
+    print("Result Loaded")
+
+except FileNotFoundError:
+
+    runner = pe(
+        data_file=hetfile,
+        par_file=parfile,
+        prior=priors,
+        detector=detector,
+        sampler="nestle",
+        sampler_kwargs={"Nlive": Nlive, "walks": 40},
+        outdir=outdir,
+        label=label,
+    )
+    
+    result = runner.result
+    
+    pickler("bilby_result.pkl", result)
+    
+  
+gridpoints = 35
+grid_size = dict()
+for p in priors.keys():
+    grid_size[p] = np.linspace(
+        np.min(result.posterior[p]), np.max(result.posterior[p]), gridpoints
+    )
+
+
+
+try:    
+    infile = open("bilby_gride.pkl",'rb')       #Try to load relevent posterior 
+    grid = pickle.load(infile)
+    infile.close()
+    print("Result Loaded")
+
+except FileNotFoundError:
+
+    grunner = pe(
+        data_file=hetfile,
+        par_file=parfile,
+        prior=priors,
+        detector=detector,
+        outdir=outdir,
+        label=label,
+        grid=True,
+        grid_kwargs={"grid_size": grid_size},
+    )
+    
+    grid = grunner.grid
+    pickler("bilby_gride.pkl", grid)
+    
+    
+    
+    
+    
+fig = result.plot_corner(save=False, parameters=injection_parameters, color="b")
+fig = corner.corner(
+    postsamples,
+    fig=fig,
+    color="r",
+    bins=50,
+    smooth=0.9,
+    quantiles=[0.16, 0.84],
+    levels=(1 - np.exp(-0.5), 1 - np.exp(-2), 1 - np.exp(-9 / 2.0)),
+    fill_contours=True,
+    hist_kwargs={"density": True},
+)
+axes = fig.get_axes()
+axidx = 0
+for p in priors.keys():
+    axes[axidx].plot(
+        grid.sample_points[p],
+        np.exp(grid.marginalize_ln_posterior(not_parameters=p) - grid.log_evidence),
+        "k--",
+    )
+    axidx += 5
+
+# custom legend
+legend_elements = [
+    Line2D([], [], color="r", label="lalapps_pulsar_parameter_estimation_nested"),
+    Line2D([], [], color="b", label="cwinpy_pe"),
+    Line2D([], [], color="k", ls="--", label="cwinpy_pe (grid)"),
+]
+font = font_manager.FontProperties(family="monospace")
+leg = axes[3].legend(
+    handles=legend_elements, loc="upper right", frameon=False, prop=font, handlelength=3
+)
+for line in leg.get_lines():
+    line.set_linewidth(1.0)
+
+fig.savefig(os.path.join(outdir, "{}_corner.png".format(label)), dpi=150)
